requirements.txt
flask==3.0.3
waitress==3.0.0
python-dotenv==1.0.1
numpy<2
pillow==10.4.0
matplotlib==3.9.2
opencv-python-headless==4.10.0.84
scikit-learn==1.5.1
pytest==8.3.2
pytest-cov==5.0.0
segmentation-models-pytorch==0.3.3
datasets>=2.20.0
huggingface-hub>=0.24.0
torch==2.4.0
torchvision==0.19.0
git+https://github.com/facebookresearch/segment-anything.git